Rename /game-v2 to /game
