+++ packages/rps/src/containers/GameFooterContainer.tsx
@@ -4,7 +4,7 @@ import GameFooter from '../components/GameFooter';
 import * as gameActions from '../redux/game/actions';
 
 import { SiteState } from '../redux/reducer';
-import { PlayingState, PlayingStateName } from '../redux/game-v2/state';
+import { PlayingState, PlayingStateName } from '../redux/game/state';
 
 function mapStateToProps(state: SiteState) {
   const localState = state.game.localState as PlayingState;
