+++ packages/rps/src/containers/__tests__/GameContainer.test.tsx
@@ -4,7 +4,7 @@ import GameContainer from '../GameContainer';
 import configureStore from 'redux-mock-store';
 import GameProposed from '../../components/GameProposedPage';
 import { SiteState } from '../../redux/reducer';
-import { localStatesA } from '../../redux/game-v2/__tests__/scenarios';
+import { localStatesA } from '../../redux/game/__tests__/scenarios';
 
 const mockStore = configureStore();
