+++ packages/rps/src/redux/game/__tests__/player-b.test.ts
