+++ packages/rps/src/redux/game/__tests__/scenarios.ts
