+++ packages/rps/src/redux/message-service/channel-updated-listener.ts
@@ -1,7 +1,7 @@
 import { take, put } from 'redux-saga/effects';
 import { buffers, eventChannel } from 'redux-saga';
 import { RPSChannelClient } from '../../utils/rps-channel-client';
-import { updateChannelState } from '../game-v2/actions';
+import { updateChannelState } from '../game/actions';
 import { ChannelState } from '../../core';
 
 export function* channelUpdatedListener() {
