+++ packages/rps/src/redux/reducer.ts
@@ -6,8 +6,8 @@ import { OpenGameState } from './open-games/state';
 import { openGamesReducer } from './open-games/reducer';
 import { overlayReducer } from './global/reducer';
 import { OverlayState } from './global/state';
-import { GameState } from './game-v2/state';
-import { gameReducer } from './game-v2/reducer';
+import { GameState } from './game/state';
+import { gameReducer } from './game/reducer';
 
 export interface SiteState {
   login: LoginState;
